@@ -1,28 +1,28 @@
 'use strict';
 
-import React, {
-  PropTypes
-} from 'react';
+import React from 'react';
+import PropTypes from 'prop-types';
 import ReactNative, {
   requireNativeComponent,
   EdgeInsetsPropType,
   StyleSheet,
   UIManager,
   View,
+  ViewPropTypes,
   NativeModules,
   Text,
   ActivityIndicator
 } from 'react-native';
+
 import resolveAssetSource from 'react-native/Libraries/Image/resolveAssetSource';
 import deprecatedPropType from 'react-native/Libraries/Utilities/deprecatedPropType';
 import invariant from 'fbjs/lib/invariant';
 import keyMirror from 'fbjs/lib/keyMirror';
-var WKWebViewManager = NativeModules.WKWebViewManager;
+const WKWebViewManager = NativeModules.WKWebViewManager;
 
 var BGWASH = 'rgba(255,255,255,0.8)';
-var RCT_WEBVIEW_REF = 'webview';
 
-var WebViewState = keyMirror({
+const WebViewState = keyMirror({
   IDLE: null,
   LOADING: null,
   ERROR: null,
@@ -47,12 +47,12 @@ type ErrorEvent = {
 
 type Event = Object;
 
-var defaultRenderLoading = () => (
+const defaultRenderLoading = () => (
   <View style={styles.loadingView}>
     <ActivityIndicator />
   </View>
 );
-var defaultRenderError = (errorDomain, errorCode, errorDesc) => (
+const defaultRenderError = (errorDomain, errorCode, errorDesc) => (
   <View style={styles.errorContainer}>
     <Text style={styles.errorTextTitle}>
       Error loading page
@@ -73,13 +73,12 @@ var defaultRenderError = (errorDomain, errorCode, errorDesc) => (
  * Renders a native WebView.
  */
 
-var WKWebView = React.createClass({
-  statics: {
-    JSNavigationScheme: JSNavigationScheme,
-    NavigationType: NavigationType,
-  },
-  propTypes: {
-    ...View.propTypes,
+class WKWebView extends React.Component {
+  static JSNavigationScheme = JSNavigationScheme;
+  static NavigationType = NavigationType;
+
+  static propTypes = {
+    ...ViewPropTypes,
 
     html: deprecatedPropType(
       PropTypes.string,
@@ -135,6 +134,18 @@ var WKWebView = React.createClass({
     ]),
 
     /**
+     * This property specifies how the safe area insets are used to modify the
+     * content area of the scroll view. The default value of this property is
+     * "never". Available on iOS 11 and later.
+     */
+    contentInsetAdjustmentBehavior: PropTypes.oneOf([
+      'automatic',
+      'scrollableAxes',
+      'never', // default
+      'always',
+    ]),
+
+    /**
      * Function that returns a view to show if there's an error.
      */
     renderError: PropTypes.func, // view to show if there's an error
@@ -163,9 +174,19 @@ var WKWebView = React.createClass({
      */
     onProgress: PropTypes.func,
     /**
-     * Receive message from webpage
+     * A function that is invoked when the webview calls `window.postMessage`.
+     * Setting this property will inject a `postMessage` global into your
+     * webview, but will still call pre-existing values of `postMessage`.
+     *
+     * `window.postMessage` accepts one argument, `data`, which will be
+     * available on the event object, `event.nativeEvent.data`. `data`
+     * must be a string.
      */
     onMessage: PropTypes.func,
+    /**
+     * Receive scroll events from view
+     */
+    onScroll: PropTypes.func,
     /**
      * @platform ios
      */
@@ -177,7 +198,7 @@ var WKWebView = React.createClass({
     onNavigationStateChange: PropTypes.func,
     scalesPageToFit: PropTypes.bool,
     startInLoadingState: PropTypes.bool,
-    style: View.propTypes.style,
+    style: ViewPropTypes.style,
     /**
      * Sets the JS to be injected when the webpage loads.
      */
@@ -204,6 +225,10 @@ var WKWebView = React.createClass({
      */
     hideKeyboardAccessoryView: PropTypes.bool,
     /**
+     * A Boolean value that determines whether pressing on a link displays a preview of the destination for the link. This props is available on devices that support 3D Touch. In iOS 10 and later, the default value is true; before that, the default value is false.
+     */
+    allowsLinkPreview: PropTypes.bool,
+    /**
      * Sets the customized user agent by using of the WKWebView
     */
     customUserAgent: PropTypes.string,
@@ -216,28 +241,27 @@ var WKWebView = React.createClass({
      * A Boolean value that sets whether diagonal scrolling is allowed.
     */
     directionalLockEnabled: PropTypes.bool,
-  },
-  getInitialState() {
-    return {
-      viewState: WebViewState.IDLE,
-      lastErrorEvent: (null: ?ErrorEvent),
-      startInLoadingState: true,
-    };
-  },
+  };
 
-  componentWillMount: function() {
+  state = {
+    viewState: WebViewState.IDLE,
+    lastErrorEvent: (null: ?ErrorEvent),
+    startInLoadingState: true,
+  };
+
+  componentWillMount() {
     if (this.props.startInLoadingState) {
-      this.setState({viewState: WebViewState.LOADING});
+      this.setState({ viewState: WebViewState.LOADING });
     }
-  },
+  }
 
   render() {
-    var otherView = null;
+    let otherView = null;
 
     if (this.state.viewState === WebViewState.LOADING) {
       otherView = (this.props.renderLoading || defaultRenderLoading)();
     } else if (this.state.viewState === WebViewState.ERROR) {
-      var errorEvent = this.state.lastErrorEvent;
+      const errorEvent = this.state.lastErrorEvent;
       invariant(
         errorEvent != null,
         'lastErrorEvent expected to be non-null'
@@ -253,21 +277,22 @@ var WKWebView = React.createClass({
       );
     }
 
-    var webViewStyles = [styles.container, styles.webView, this.props.style];
+    const webViewStyles = [styles.container, styles.webView, this.props.style];
     if (this.state.viewState === WebViewState.LOADING ||
       this.state.viewState === WebViewState.ERROR) {
       // if we're in either LOADING or ERROR states, don't show the webView
       webViewStyles.push(styles.hidden);
     }
 
-    var onShouldStartLoadWithRequest = this.props.onShouldStartLoadWithRequest && ((event: Event) => {
-      var shouldStart = this.props.onShouldStartLoadWithRequest &&
+    const onShouldStartLoadWithRequest = this.props.onShouldStartLoadWithRequest && ((event: Event) => {
+      const shouldStart = this.props.onShouldStartLoadWithRequest &&
         this.props.onShouldStartLoadWithRequest(event.nativeEvent);
       WKWebViewManager.startLoadWithResult(!!shouldStart, event.nativeEvent.lockIdentifier);
     });
 
+    let source = {};
     if (this.props.source && typeof this.props.source == 'object') {
-      var source = Object.assign({}, this.props.source, { 
+      source = Object.assign({}, this.props.source, {
         sendCookies: this.props.sendCookies,
         customUserAgent: this.props.customUserAgent || this.props.userAgent
       });
@@ -279,11 +304,14 @@ var WKWebView = React.createClass({
       source.uri = this.props.url;
     }
 
-    var webView =
+    const messagingEnabled = typeof this.props.onMessage === 'function';
+
+    const webView =
       <RCTWKWebView
-        ref={RCT_WEBVIEW_REF}
+        ref={ref => { this.webview = ref; }}
         key="webViewKey"
         style={webViewStyles}
+        contentInsetAdjustmentBehavior={this.props.contentInsetAdjustmentBehavior}
         source={resolveAssetSource(source)}
         injectedJavaScript={this.props.injectedJavaScript}
         bounces={this.props.bounces}
@@ -293,11 +321,14 @@ var WKWebView = React.createClass({
         automaticallyAdjustContentInsets={this.props.automaticallyAdjustContentInsets}
         openNewWindowInWebView={this.props.openNewWindowInWebView}
         hideKeyboardAccessoryView={this.props.hideKeyboardAccessoryView}
+        allowsLinkPreview={this.props.allowsLinkPreview}
         onLoadingStart={this._onLoadingStart}
         onLoadingFinish={this._onLoadingFinish}
         onLoadingError={this._onLoadingError}
+        messagingEnabled={messagingEnabled}
         onProgress={this._onProgress}
         onMessage={this._onMessage}
+        onScroll={this._onScroll}
         onShouldStartLoadWithRequest={onShouldStartLoadWithRequest}
         pagingEnabled={this.props.pagingEnabled}
         directionalLockEnabled={this.props.directionalLockEnabled}
@@ -309,96 +340,114 @@ var WKWebView = React.createClass({
         {otherView}
       </View>
     );
-  },
+  }
 
   /**
    * Go forward one page in the webview's history.
    */
-  goForward: function() {
+  goForward = () => {
     UIManager.dispatchViewManagerCommand(
       this.getWebViewHandle(),
       UIManager.RCTWKWebView.Commands.goForward,
       null
     );
-  },
+  };
 
   /**
    * Go back one page in the webview's history.
    */
-  goBack: function() {
+  goBack = () => {
     UIManager.dispatchViewManagerCommand(
       this.getWebViewHandle(),
       UIManager.RCTWKWebView.Commands.goBack,
       null
     );
-  },
+  };
 
   /**
    * Indicating whether there is a back item in the back-forward list that can be navigated to
    */
-  canGoBack: function() {
+  canGoBack = () => {
     return WKWebViewManager.canGoBack(this.getWebViewHandle());
-  },
+  };
 
   /**
    * Indicating whether there is a forward item in the back-forward list that can be navigated to
    */
-  canGoForward: function() {
+  canGoForward = () => {
     return WKWebViewManager.canGoForward(this.getWebViewHandle());
-  },
+  };
 
   /**
    * Reloads the current page.
    */
-  reload: function() {
+  reload = () => {
     UIManager.dispatchViewManagerCommand(
       this.getWebViewHandle(),
       UIManager.RCTWKWebView.Commands.reload,
       null
     );
-  },
+  };
 
   /**
    * Stop loading the current page.
    */
-  stopLoading: function() {
+  stopLoading = () => {
     UIManager.dispatchViewManagerCommand(
       this.getWebViewHandle(),
       UIManager.RCTWKWebView.Commands.stopLoading,
       null
     )
-  },
+  };
 
-  evaluateJavaScript: function(js) {
+  /**
+   * Posts a message to the web view, which will emit a `message` event.
+   * Accepts one argument, `data`, which must be a string.
+   *
+   * In your webview, you'll need to something like the following.
+   *
+   * ```js
+   * document.addEventListener('message', e => { document.title = e.data; });
+   * ```
+   */
+  postMessage = (data) => {
+    UIManager.dispatchViewManagerCommand(
+      this.getWebViewHandle(),
+      UIManager.RCTWKWebView.Commands.postMessage,
+      [String(data)]
+    );
+  };
+
+  evaluateJavaScript = (js) => {
     return WKWebViewManager.evaluateJavaScript(this.getWebViewHandle(), js);
-  },
+  };
 
   /**
    * We return an event with a bunch of fields including:
    *  url, title, loading, canGoBack, canGoForward
    */
-  _updateNavigationState: function(event: Event) {
+  _updateNavigationState = (event: Event) => {
     if (this.props.onNavigationStateChange) {
       this.props.onNavigationStateChange(event.nativeEvent);
     }
-  },
+  };
 
   /**
    * Returns the native webview node.
    */
-  getWebViewHandle: function(): any {
-    return ReactNative.findNodeHandle(this.refs[RCT_WEBVIEW_REF]);
-  },
+  getWebViewHandle = (): any => {
+    return ReactNative.findNodeHandle(this.webview);
+  };
 
-  _onLoadingStart: function(event: Event) {
-    var onLoadStart = this.props.onLoadStart;
+  _onLoadingStart = (event: Event) => {
+    const onLoadStart = this.props.onLoadStart;
     onLoadStart && onLoadStart(event);
     this._updateNavigationState(event);
-  },
+  };
 
-  _onLoadingError: function(event: Event) {
+  _onLoadingError = (event: Event) => {
     event.persist(); // persist this event because we need to store it
-    var {onError, onLoadEnd} = this.props;
+    const { onError, onLoadEnd } = this.props;
     onError && onError(event);
     onLoadEnd && onLoadEnd(event);
     console.warn('Encountered an error loading page', event.nativeEvent);
@@ -407,38 +456,45 @@ var WKWebView = React.createClass({
       lastErrorEvent: event.nativeEvent,
       viewState: WebViewState.ERROR
     });
-  },
+  };
 
-  _onLoadingFinish: function(event: Event) {
-    var {onLoad, onLoadEnd} = this.props;
+  _onLoadingFinish = (event: Event) => {
+    const { onLoad, onLoadEnd } = this.props;
     onLoad && onLoad(event);
     onLoadEnd && onLoadEnd(event);
     this.setState({
       viewState: WebViewState.IDLE,
     });
     this._updateNavigationState(event);
-  },
+  };
 
-  _onProgress(event: Event) {
-    var onProgress = this.props.onProgress;
+  _onProgress = (event: Event) => {
+    const onProgress = this.props.onProgress;
     onProgress && onProgress(event.nativeEvent.progress);
-  },
+  };
 
-  _onMessage(event: Event) {
-    var onMessage = this.props.onMessage;
-    onMessage && onMessage(event.nativeEvent);
-  }
-});
+  _onMessage = (event: Event) => {
+    var { onMessage } = this.props;
+    onMessage && onMessage(event);
+  };
 
-var RCTWKWebView = requireNativeComponent('RCTWKWebView', WKWebView, {
+  _onScroll = (event: Event) => {
+    const onScroll = this.props.onScroll;
+    onScroll && onScroll(event.nativeEvent);
+  };
+}
+
+const RCTWKWebView = requireNativeComponent('RCTWKWebView', WKWebView, {
   nativeOnly: {
     onLoadingStart: true,
     onLoadingError: true,
     onLoadingFinish: true,
+    onMessage: true,
+    messagingEnabled: PropTypes.bool,
   }
 });
 
-var styles = StyleSheet.create({
+const styles = StyleSheet.create({
   container: {
     flex: 1,
   },
